% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.extracted.R
\name{c.extracted}
\alias{c.extracted}
\title{Combines a extracted Object}
\usage{
\method{c}{extracted}(x, ...)
}
\arguments{
\item{x}{The extracted object}

\item{\ldots}{ignored}
}
\description{
Combines a extracted object
}
