% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat.R
\name{cheat}
\alias{cheat}
\title{A Cheat Sheet of Common Regex Task Chunks}
\usage{
cheat(dictionary = qdapRegex::regex_cheat, print = TRUE)
}
\arguments{
\item{dictionary}{A dictionary of cheat terms.  Default is 
\code{\link[qdapRegex]{regex_cheat}}.}

\item{print}{logical.  If \code{TRUE} the left justified output is printed 
to the console.}
}
\value{
Prints a cheat sheet of common regex tasks such as lookaheads.  
Invisibly returns \code{\link[qdapRegex]{regex_cheat}}.
}
\description{
Print a cheat sheet of common regex task chunks.  \code{cheat} prints a left 
justified version of \code{\link[qdapRegex]{regex_cheat}}.
}
\examples{
cheat()
}
\seealso{
\code{\link[qdapRegex]{regex_cheat}}
}
