% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grab}
\alias{grab}
\title{Grab Regular Expressions from Dictionaries}
\usage{
grab(pattern, dictionary = getOption("regex.library"))
}
\arguments{
\item{pattern}{A character string starting with \code{"@"} and a regular
expression from a pkg{qdapRegex} dictionary.}

\item{dictionary}{A dictionary of canned regular expressions to search within.}
}
\value{
Returns a single string regular expressionfrom one of the pkg{qdapRegex}
dictionaries.
}
\description{
A conveinence function to
}
\examples{
grab("@rm_white")
\dontrun{
## Throws an error
grab("@foo")
}
}
\keyword{get}
\keyword{grab}

