% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{print.explain}
\alias{print.explain}
\title{Prints a explain object}
\usage{
\method{print}{explain}(x, ...)
}
\arguments{
\item{x}{The explain object}

\item{\ldots}{ignored}
}
\description{
Prints a explain object
}
