% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.extracted.R
\name{print.extracted}
\alias{print.extracted}
\title{Prints a extracted Object}
\usage{
\method{print}{extracted}(x, ...)
}
\arguments{
\item{x}{The \code{extracted} object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a \code{extracted} object
}
