% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.regexr.R
\name{print.regexr}
\alias{print.regexr}
\title{Prints a regexr Object}
\usage{
\method{print}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a \code{regexr} object
}
