% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{qdapRegex}
\alias{package-qdapRegex}
\alias{qdapRegex}
\alias{qdapRegex-package}
\title{qdapRegex: Regular expression removal/extraction/replacement tools for the \pkg{qdap} package}
\description{
This package is a collection of regex tools associated with the
\pkg{qdap} package that may be useful outside of the context of discourse
analysis.  Tools include removal/extraction/replacement of dabbreviations,
dates, dollar amounts, email addresses, hash tags, numbers, percentages,
person tags, phone numbers, times, and zip codes.
}

