% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{RE}
\alias{RE}
\title{Canned Regular Expressions}
\format{A list with 14 elements}
\usage{
data(RE)
}
\description{
A dataset containing a list canned regular expressions used in various
functions within the \pkg{qdapRegex} package.
}
\keyword{datasets}

