% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{regex_cheat}
\alias{regex_cheat}
\title{A Cheatsheet of Common Regex Task Chunks}
\format{A data frame with 6 rows and 3 variables}
\usage{
data(regex_cheat)
}
\description{
A dataset containing the regex chunk name, the regex string, and a
description of what the chunk does.
}
\details{
\itemize{
  \item Name. The name of the regex chunk.
  \item Regex. The regex chunk.
  \item What it Does. Description of what the regex chunk does.
}
}
\references{
\url{http://www.rexegg.com}
}
\keyword{datasets}

