% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{regex_usa}
\alias{regex_usa}
\title{Canned Regular Expressions (United States of America)}
\format{A list with 21 elements}
\usage{
data(regex_usa)
}
\description{
A dataset containing a list U.S. specific, canned regular expressions for use
in various functions within the \pkg{qdapRegex} package.
}
\keyword{datasets}

