% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_between}
\alias{rm_between}
\title{Remove/Replace/Extract Strings Between 2 Markers}
\usage{
rm_between(text.var, left, right, trim = TRUE, clean = TRUE,
  replacement = "", extract = FALSE, include.markers = FALSE,
  merge = TRUE)
}
\arguments{
\item{text.var}{The text variable.}

\item{left}{A vector of character or numeric symbols as the left edge to
extract.}

\item{right}{A vector of character or numeric symbols as the right edge to
extract.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the strings are extracted into a
list of vectors.}

\item{include.markers}{logical.  If \code{TRUE} and \code{extract = TRUE} returns
the markers (left/right) and the text between.}

\item{merge}{logical.  If \code{TRUE} the results of each bracket type will
be merged by string.  \code{FALSE} returns a named list of lists of vectors
of markered text per marker type.}
}
\value{
Returns a character string with markers removed.
}
\description{
Remove/replace/extract strings bounded between a left and right marker.
}
\examples{
x <- c("Where is the /big dog#?",
    "I think he's @arunning@b with /little cat#.")

rm_between(x, "@a", "@b")
rm_between(x, "@a", "@b", extract=TRUE)
rm_between(x, c("/", "@a"), c("#", "@b"))
rm_between(x, c("/", "@a"), c("#", "@b"), extract=TRUE)

x2 <- c("Where is the L1big dogL2?",
    "I think he's 98running99 with L1little catL2.")
rm_between(x2, c("L1", 98), c("L2", 99))
rm_between(x2, c("L1", 98), c("L2", 99), extract=TRUE)

state <- c("Computer is fun. Not too fun.", "No it's not, it's dumb.",
    "What should we do?", "You liar, it stinks!", "I am telling the truth!",
    "How can we be certain?", "There is no way.", "I distrust you.",
    "What are you talking about?", "Shall we move on?  Good then.",
    "I'm hungry.  Let's eat.  You already?")

rm_between(x2, c("is", "we"), c("too", "on"))
}
\seealso{
\code{\link[base]{gsub}},
\code{\link[qdapRegex]{rm_bracket}}
}

