% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_city_state}
\alias{rm_city_state}
\title{Remove/Replace/Extract City & State}
\usage{
rm_city_state(text.var, trim = TRUE, clean = TRUE,
  pattern = qdapRegex::RE[["rm_city_state"]], replacement = "",
  extract = FALSE, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the city & state are extracted into a
list of vectors.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with city & state removed.
}
\description{
Remove/replace/extract city (single lower case word or multiple consecutive capitalized words before a comma and state) & state (2 consecutive capital letters) from a string in the form of XX/XX/XXXX,
XX/XX/XX, XX-XX-XXXX, XX-XX-XX, XX.XX.XXXX, or XX.XX.XX.
}
\examples{
x <- paste0("I went to Washington Heights, NY for food! ",
   "It's in West ven,PA, near Bolly Bolly Bolly, CA!",
   "I like Movies, PG13")
rm_city_state(x)
rm_city_state(x, extract=TRUE)
}
\seealso{
\code{\link[base]{gsub}}
}
\keyword{date}

