% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rm_default.R
\name{rm_default}
\alias{rm_default}
\title{Remove/Replace/Extract Template}
\usage{
rm_default(text.var, trim = !extract, clean = TRUE, pattern,
  replacement = "", extract = FALSE,
  dictionary = getOption("regex.library"), ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the strings are extracted into a list
of vectors.}

\item{dictionary}{A dictionary of canned regular expressions to search within
if \code{pattern} begins with \code{"@rm_"}.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with substring removed.
}
\description{
Remove/replace/extract substring from a string.  This is the template used by
other \pkg{qdapRegex} \code{rm_XXX} functions.
}
\examples{
## Built in regex dictionary
rm_default("I live in Buffalo, NY 14217", pattern="@rm_city_state_zip")

## User defined regular expression
pat <- "(\\\\s*([A-Z][\\\\w-]*)+),\\\\s([A-Z]{2})\\\\s(?<!\\\\d)\\\\d{5}(?:[ -]\\\\d{4})?\\\\b"
rm_default("I live in Buffalo, NY 14217", pattern=pat)
}
\seealso{
\code{\link[qdapRegex]{rm_}},
\code{\link[base]{gsub}},
\code{\link[stringi]{stri_extract_all_regex}}
}
\keyword{extract}
\keyword{sub}

