% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rm_phone}
\alias{rm_phone}
\title{Remove/Replace/Extract Phone Numbers}
\usage{
rm_phone(text.var, trim = TRUE, clean = TRUE, pattern = "@rm_phone",
  replacement = "", extract = FALSE,
  dictionary = getOption("regex.library"), ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.  Default, \code{@rm_phone} uses the
\code{rm_phone} regex from the regular expression dictionary from
the \code{dictionary} argument.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the phone numbers are extracted into a
list of vectors.}

\item{dictionary}{A dictionary of canned regular expressions to search within
if \code{pattern} begins with \code{"@rm_"}.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with phone numbers removed.
}
\description{
Remove/replace/extract phone numbers from a string.
}
\examples{
x <- c(" Mr. Bean bought 2 tickets 2-613-213-4567 or 5555555555 call either one",
  "43 Butter Rd, Brossard QC K0A 3P0 - 613 213 4567",
  "Please contact Mr. Bean (613)2134567",
  "1.575.555.5555 is his #1 number",
  "7164347566",
  "I like 1234567 dogs"
)

rm_phone(x)
rm_phone(x, extract=TRUE)
}
\author{
\href{http://stackoverflow.com/}{stackoverflow's} Marius and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
The phone regular expression was taken from:
\url{http://stackoverflow.com/a/21008254/1000343}
}
\seealso{
\code{\link[base]{gsub}}
}
\keyword{phone}
\keyword{telephone}

