% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rm_simile_as.R
\name{rm_simile_as}
\alias{rm_simile_as}
\title{Remove/Replace/Extract Simile (as ___  as)}
\usage{
rm_simile_as(text.var, trim = !extract, clean = TRUE,
  pattern = "@rm_simile_as", replacement = "", extract = FALSE,
  dictionary = getOption("regex.library"), ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{trim}{logical.  If \code{TRUE} removes leading and trailing white
spaces.}

\item{clean}{trim logical.  If \code{TRUE} extra white spaces and escaped
character will be removed.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given
character vector.}

\item{replacement}{Replacement for matched \code{pattern}.}

\item{extract}{logical.  If \code{TRUE} the person tags are extracted into a
list of vectors.}

\item{dictionary}{A dictionary of canned regular expressions to search within
if \code{pattern} begins with \code{"@rm_"}.}

\item{\dots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character string with person tags removed.
}
\description{
Remove/replace/extract simile (as ___  as) from a string.  Useful more so for
the ability to split on these points.
}
\examples{
sims_as <- c(
    "John is as big as a horse and chicken.",
    "He's about as entertaining as watching paint dry",
    "This is as good as it gets",
    "It's not as good"
)

rm_simile_as(sims_as, replacement="--")
rm_simile_as(sims_as, replacement="--", extract = TRUE)

\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr)

sims_as[grepl(grab("@rm_simile_as"), sims_as)] \%>\%
    rm_simile_as(replacement="<<SIMILE>> \\\\2 <<SIMILE>>") \%>\%
    strsplit("\\\\s*<<SIMILE>>\\\\s*") \%>\%
    do.call(rbind.data.frame, .) \%>\%
    setNames(c("Description", paste0("Noun_Phrase_", 1:2))[c(2, 1, 3)])
}
}
\seealso{
\code{\link[base]{gsub}},
\code{\link[stringi]{stri_extract_all_regex}}

Other rm_simile.functions: \code{\link{rm_simile_like}}
}
\keyword{person}
\keyword{tag}
\keyword{twitter}

