% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S}
\alias{S}
\title{Use C-style String Formatting Commands}
\usage{
S(x, ...)
}
\arguments{
\item{x}{A single string containing \code{"\%s"}.}

\item{\ldots}{A vector of substitutions equal in length to the number of 
\code{"\%s"} in \code{x} or of length one (if length one \ldots will be recycled).}
}
\value{
Returns a string with \code{"\%s"} replaced.
}
\description{
Convenience wrapper for \code{\link[base]{sprintf}} that allows recycling of 
\ldots of length one.
}
\examples{
S("@after_", "the", "the")
# Recycle
S("@after_", "the")
S("@rm_between", "LEFT", "RIGHT")
}
\seealso{
\code{\link[base]{sprintf}}
}
