% Generated by roxygen2 (4.0.2): do not edit by hand
\name{explain}
\alias{explain}
\title{Visualize Regular Expressions}
\usage{
explain(pattern, dictionary = getOption("regex.library"))
}
\arguments{
\item{pattern}{A character string containing a regular expression or a
character string starting with \code{"@"} that is a regular
expression from a \pkg{qdapRegex} dictionary.}

\item{dictionary}{A dictionary of canned regular expressions to search within.}
}
\value{
Attemps to open the url to the visual representation provided by
\url{http://www.regexper.com} and invisibly returns the url string.
}
\description{
Visualize regular expressions using \url{http://www.regexper.com}.
}
\examples{
\dontrun{
explain("\\\\s*foo[A-Z]\\\\d{2,3}")
explain("@rm_time")
}
}
\author{
Ananda Mahto and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/27489977/1000343}
}
\seealso{
\url{http://www.regexper.com}
}
\keyword{explain}
\keyword{regex}

