% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qdapRegex-package.R
\docType{package}
\name{qdapRegex}
\alias{package-qdapRegex}
\alias{qdapRegex}
\alias{qdapRegex-package}
\title{qdapRegex: Regular Expression Removal, Extraction, & Replacement Tools for the \pkg{qdap} Package}
\description{
\pkg{qdapRegex} is a collection of regular expression tools associated with
the \pkg{qdap} package that may be useful outside of the context of discourse
analysis.  Tools include removal/extraction/replacement of abbreviations,
dates, dollar amounts, email addresses, hash tags, numbers, percentages,
citations, person tags, phone numbers, times, and zip codes.
}
\details{
The \pkg{qdapRegex} package does not aim to compete with string manipulation
packages such as
\href{http://cran.r-project.org/package=stringr}{\pkg{stringr}}
or \href{http://cran.r-project.org/package=stringi}{\pkg{stringi}}
but is meant to provide access to canned, common regular expression patterns
that can be used within \pkg{qdapRegex}, with \pkg{R}'s own regular
expression functions, or add on string manipulation packages such as
\code{stringr} and \code{stringi}.
}

