% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{qdapRegex}
\alias{package-qdapRegex}
\alias{qdapRegex}
\alias{qdapRegex-package}
\title{qdapRegex: Regular expression removal/extraction/replacement tools for the \pkg{qdap} package}
\description{
This package is a collection of  regular expression tools associated with the
\pkg{qdap} package that may be useful outside of the context of discourse
analysis.  Tools include removal/extraction/replacement of abbreviations,
dates, dollar amounts, email addresses, hash tags, numbers, percentages,
person tags, phone numbers, times, and zip codes.
}
\details{
The \pkg{qdapRegex} package does not aim to compete with string manipulation
packages such as \pkg{stringr} or \pkg{stringi} but is meant to provide
access to canned, common regular expression patterns that can be used within
\pkg{qdapRegex}, with \pkg{R}'s on regular expression functions, or add on
string manipulation packages.
}

