% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdapRegex-package.R
\docType{package}
\name{qdapRegex}
\alias{qdapRegex-package}
\alias{qdapRegex}
\alias{package-qdapRegex}
\title{qdapRegex: Regular Expression Removal, Extraction, & Replacement Tools for the \pkg{qdap} Package}
\description{
\pkg{qdapRegex} is a collection of regular expression tools associated with 
the \pkg{qdap} package that may be useful outside of the context of discourse
analysis.  Tools include removal/extraction/replacement of abbreviations, 
dates, dollar amounts, email addresses, hash tags, numbers, percentages, 
citations, person tags, phone numbers, times, and zip codes.
}
\details{
The \pkg{qdapRegex} package does not aim to compete with string manipulation 
packages such as 
\href{https://CRAN.R-project.org/package=stringr}{\pkg{stringr}} 
or \href{https://CRAN.R-project.org/package=stringi}{\pkg{stringi}} 
but is meant to provide access to canned, common regular expression patterns 
that can be used within \pkg{qdapRegex}, with \pkg{R}'s own regular 
expression functions, or add on string manipulation packages such as 
\code{stringr} and \code{stringi}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trinker/qdapRegex}
  \item Report bugs at \url{https://github.com/trinker/qdapRegex/issues}
}

}
\author{
\strong{Maintainer}: Tyler Rinker \email{tyler.rinker@gmail.com}

Other contributors:
\itemize{
  \item Jason Gray [contributor]
}

}
